refactor SDK
