+++ screenpipe-js/common/types.ts
