--- screenpipe-js/next.ts
@@ -1,27 +1,0 @@
-// next.ts
-// Types and utilities shared between browser and Node environments
-export * from "./types";
-
-// Helper functions that work in both environments
-export function toCamelCase(str: string): string {
-  return str.replace(/([-_][a-z])/g, (group) =>
-    group.toUpperCase().replace("-", "").replace("_", "")
-  );
-}
-
-export function toSnakeCase(str: string): string {
-  return str.replace(/[A-Z]/g, (letter) => `_${letter.toLowerCase()}`);
-}
-
-export function convertToCamelCase(obj: any): any {
-  if (Array.isArray(obj)) {
-    return obj.map(convertToCamelCase);
-  } else if (obj !== null && typeof obj === "object") {
-    return Object.keys(obj).reduce((result, key) => {
-      const camelKey = toCamelCase(key);
-      result[camelKey] = convertToCamelCase(obj[key]);
-      return result;
-    }, {} as any);
-  }
-  return obj;
-}
