+++ screenpipe-js/node-sdk/.gitignore
