+++ screenpipe-js/node-sdk/README.md
@@ -1,0 +1,15 @@
+# test-screenpipe-node-sdk
+
+To install dependencies:
+
+```bash
+bun install
+```
+
+To run:
+
+```bash
+bun run ./dist/main.js
+```
+
+This project was created using `bun init` in bun v1.1.42. [Bun](https://bun.sh) is a fast all-in-one JavaScript runtime.
