+++ screenpipe-js/node-sdk/package.json
@@ -1,0 +1,42 @@
+{
+  "name": "@screenpipe/node",
+  "version": "0.1.12",
+  "type": "module",
+  "main": "./dist/index.cjs",
+  "module": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "tsup ./src --format esm,cjs --dts --clean",
+    "test": "bun test",
+    "prepublishOnly": "npm run build"
+  },
+  "devDependencies": {
+    "@types/bun": "^1.1.14",
+    "@types/node": "^22.10.4",
+    "bun-types": "latest",
+    "tsup": "^8.0.0"
+  },
+  "dependencies": {
+    "@scarf/scarf": "^1.4.0",
+    "@types/express": "^5.0.0",
+    "@types/node-cron": "^3.0.11",
+    "@types/uuid": "^10.0.0",
+    "express": "^4.21.1",
+    "node-cron": "^3.0.3"
+  },
+  "peerDependencies": {
+    "typescript": "^5.0.0"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "typesVersions": {
+    "*": {
+      "node": ["./dist/node.d.ts"],
+      "browser": ["./dist/browser.d.ts"]
+    }
+  }
+}
