+++ screenpipe-js/node-sdk/src/Task.ts
@@ -1,0 +1,69 @@
+import cron from "node-cron";
+
+export class Task {
+  private _name: string;
+  private _interval: string | number;
+  private _time: string | null = null;
+  private _handler: (() => Promise<void>) | null = null;
+  private _cronTask: any | null = null;
+
+  constructor(name: string) {
+    this._name = name;
+    this._interval = 0;
+  }
+
+  every(interval: string | number): Task {
+    this._interval = interval;
+    return this;
+  }
+
+  at(time: string): Task {
+    this._time = time;
+    return this;
+  }
+
+  do(handler: () => Promise<void>): Task {
+    this._handler = handler;
+    return this;
+  }
+
+  async schedule(): Promise<void> {
+    if (!this._handler) {
+      throw new Error(`No handler defined for task: ${this._name}`);
+    }
+    const cronExpression = this.toCronExpression();
+
+    this._cronTask = cron.schedule(cronExpression, this._handler, {
+      name: this._name,
+    });
+  }
+
+  stop(): void {
+    return this._cronTask?.stop();
+  }
+
+  private toCronExpression(): string {
+    if (typeof this._interval === "number") {
+      const minutes = Math.floor(this._interval / 60000);
+      return `*/${minutes} * * * *`;
+    }
+
+    const [value, unit] = this._interval.split(" ");
+    switch (unit) {
+      case "second":
+      case "seconds":
+        return `*/${value} * * * * *`;
+      case "minute":
+      case "minutes":
+        return `*/${value} * * * *`;
+      case "hour":
+      case "hours":
+        return `0 */${value} * * *`;
+      case "day":
+      case "days":
+        return `0 0 */${value} * *`;
+      default:
+        throw new Error(`Unsupported interval unit: ${unit}`);
+    }
+  }
+}
