--- screenpipe-js/package.json
@@ -1,60 +1,0 @@
-{
-    "name": "@screenpipe/js",
-    "version": "0.1.11",
-    "type": "module",
-    "main": "./dist/main.cjs",
-    "module": "./dist/main.js",
-    "types": "./dist/main.d.ts",
-    "files": [
-        "dist"
-    ],
-    "exports": {
-        ".": {
-            "types": "./dist/main.d.ts",
-            "browser": "./dist/browser.js",
-            "node": "./dist/node.js",
-            "import": "./dist/main.js",
-            "require": "./dist/main.cjs",
-            "default": "./dist/browser.js"
-        },
-        "./node": {
-            "types": "./dist/node.d.ts",
-            "import": "./dist/node.js",
-            "require": "./dist/node.cjs"
-        },
-        "./browser": {
-            "types": "./dist/browser.d.ts",
-            "import": "./dist/browser.js",
-            "require": "./dist/browser.cjs"
-        }
-    },
-    "scripts": {
-        "build": "tsup main.ts browser.ts node.ts next.ts types.ts --format esm,cjs --dts --clean",
-        "test": "bun test",
-        "prepublishOnly": "npm run build"
-    },
-    "devDependencies": {
-        "@types/node": "^22.10.2",
-        "bun-types": "latest",
-        "tsup": "^8.0.0"
-    },
-    "dependencies": {
-        "@scarf/scarf": "^1.4.0",
-        "@types/express": "^5.0.0",
-        "@types/node-cron": "^3.0.11",
-        "express": "^4.21.1",
-        "node-cron": "^3.0.3"
-    },
-    "peerDependencies": {
-        "typescript": "^5.0.0"
-    },
-    "publishConfig": {
-        "access": "public"
-    },
-    "typesVersions": {
-        "*": {
-            "node": ["./dist/node.d.ts"],
-            "browser": ["./dist/browser.d.ts"]
-        }
-    }
-}
