--- screenpipe-js/tsconfig.json
@@ -1,26 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2020",
-    "module": "ESNext",
-    "moduleResolution": "bundler",
-    "types": [
-      "bun-types",
-      "node"
-    ],
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "strict": true,
-    "declaration": true,
-    "declarationDir": "./dist",
-    "outDir": "dist"
-  },
-  "include": [
-    "*.ts",
-    "index.d.ts"
-  ],
-  "exclude": [
-    "node_modules",
-    "dist",
-    "create-pipe"
-  ]
-}
